requirements.txt
discord.py
python-dotenv
feedparser
flask